+++ themes-default/slim/src/components/config-post-processing.vue
@@ -34,7 +34,7 @@
 
                         <div class="row component-group">
                             <div class="component-group-desc col-xs-12 col-md-2">
-                                <h3>Automated Download Handling</h3>
+                                <a name="automated-download-handling" /><h3>Automated Download Handling</h3>
                                 <p>Check clients directly through api's for completed or failed downloads.</p>
                                 <p>The download handler will periodically connect to the nzb or torrent clients and check for completed and failed downloads.</p>
                             </div>
@@ -69,6 +69,10 @@
                                         <p>Setting the ratio to 0, will have it perform the action directly after postprocessing.)</p>
                                     </config-template>
 
+                                    <config-template label-for="default_client_path" label="Default client path">
+                                        <file-browser id="default_client_path" name="default_client_path" title="Select client download location" :initial-dir="postprocessing.defaultClientPath" @update="postprocessing.defaultClientPath = $event" />
+                                        <span class="clear-left">To prevent postprocessing from deleting your (root) download location, select the location to protect it from removal.</span>
+                                    </config-template>
                                 </fieldset>
                             </div> <!-- end of col -->
                         </div> <!-- end of row -->
@@ -88,11 +92,33 @@
                                     </config-template>
 
                                     <config-template label-for="processing_method" label="Processing Method">
-                                        <select id="naming_multi_ep" name="naming_multi_ep" v-model="postprocessing.processMethod" class="form-control input-sm">
+                                        <select id="processing_method" name="processing_method" v-model="postprocessing.processMethod" class="form-control input-sm">
                                             <option :value="option.value" v-for="option in processMethods" :key="option.value">{{ option.text }}</option>
                                         </select>
                                         <span>What method should be used to put files into the library?</span>
                                         <p><b>Note:</b> If you keep seeding torrents after they finish, please avoid the 'move' processing method to prevent errors.</p>
+                                        <p v-if="postprocessing.processMethod == 'reflink'">To use reference linking, the <app-link href="http://www.dereferer.org/?https://pypi.python.org/pypi/reflink/0.1.4">reflink package</app-link> needs to be installed.</p>
+                                    </config-template>
+
+                                    <config-toggle-slider v-model="postprocessing.specificPostProcessing" label="Specific postprocessing methods" id="specific_post_processing">
+                                        <span>Enable this option if you want to use different processing methods (copy, move, etc..) for torrent and nzb downloads.</span>
+                                        <p><b>Note:</b>This option is only used by the <a href="config/postProcessing/#automated-download-handling">Automated Download Handling</a> option</p>
+                                    </config-toggle-slider>
+
+                                    <config-template v-if="postprocessing.specificPostProcessing" label-for="processing_method_torrent" label="Processing Method Torrent">
+                                        <select id="processing_method_torrent" name="processing_method_torrent" v-model="postprocessing.processMethodTorrent" class="form-control input-sm">
+                                            <option :value="option.value" v-for="option in processMethods" :key="option.value">{{ option.text }}</option>
+                                        </select>
+                                        <span>What method should be used to put files into the library?</span>
+                                        <p><b>Note:</b> If you keep seeding torrents after they finish, please avoid the 'move' processing method to prevent errors.</p>
+                                        <p v-if="postprocessing.processMethod == 'reflink'">To use reference linking, the <app-link href="http://www.dereferer.org/?https://pypi.python.org/pypi/reflink/0.1.4">reflink package</app-link> needs to be installed.</p>
+                                    </config-template>
+
+                                    <config-template v-if="postprocessing.specificPostProcessing" label-for="processing_method_nzb" label="Processing Method Nzb">
+                                        <select id="processing_method_nzb" name="processing_method_nzb" v-model="postprocessing.processMethodNzb" class="form-control input-sm">
+                                            <option :value="option.value" v-for="option in processMethods" :key="option.value">{{ option.text }}</option>
+                                        </select>
+                                        <span>What method should be used to put files into the library?</span>
                                         <p v-if="postprocessing.processMethod == 'reflink'">To use reference linking, the <app-link href="http://www.dereferer.org/?https://pypi.python.org/pypi/reflink/0.1.4">reflink package</app-link> needs to be installed.</p>
                                     </config-template>