+++ themes-default/slim/src/components/manage-episode-status.vue
@@ -2,7 +2,7 @@
     <div id="episode-status">
         <div id="select-status">
             <span>Manage episodes with status</span>
-            <select :disabled="manageStatus" name="whichStatus" v-model="selectedStatus" class="form-control form-control-inline input-sm">
+            <select :disabled="manageStatus" v-model="selectedStatus" class="form-control form-control-inline input-sm">
                 <option v-for="status in availableStatus" :value="status.value" :key="status.value">{{status.text}}</option>
             </select>
 
@@ -178,6 +178,13 @@ export default {
         clearPage() {
             this.manageStatus = null;
             this.data = [];
+        }
+    },
+    mounted() {
+        const { status } = this.$route.query;
+        if (status) {
+            this.selectedStatus = Number(status);
+            this.getEpisodes();
         }
     }
 };