+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 ## Unreleased
 
 #### New Features
+- Added separate configs for the process methods (copy, move, etc) for torrent and nzb. (only usuable with the download handler) ([9932](https://github.com/pymedusa/Medusa/pull/9932))
+- Added setting for the default client path that will be protected (can't be deleted) during post-processing ([9954](https://github.com/pymedusa/Medusa/pull/9954))
 
 #### Improvements
 
 #### Fixes
+- Correctly delete folders with files for move method or if explicitly wanted ([9950](https://github.com/pymedusa/Medusa/pull/9950))
+- Fixed link to the overview of snatched episodes at the bottom of the pages ([9954](https://github.com/pymedusa/Medusa/pull/9954))
 
 -----
 
@@ -16,7 +20,7 @@
 
 #### Fixes
 - Fix prowlarr provider id's being obfuscated in logs because of a bad log level. ([9857](https://github.com/pymedusa/Medusa/pull/9857))
-- Fix postprocessing specials. ([9812](https://github.com/pymedusa/Medusa/pull/9812)) 
+- Fix postprocessing specials. ([9812](https://github.com/pymedusa/Medusa/pull/9812))
 - Fix storing a negative value in the UI as a search delay value ([9822](https://github.com/pymedusa/Medusa/pull/9822))
 
 -----